requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.3.2
pymongo==4.5.0
pydantic==2.7.1
python-dotenv
